feat: add indexing each test step
